+++ contracts/RETypes.sol
@@ -2,26 +2,28 @@
 pragma solidity ^0.8.9;
 
 library RETypes {
-
     struct ActionRuntimeParams {
-        uint triggerData; 
-        uint totalCollateralAmount; 
+        uint256 triggerData;
+        uint256 totalCollateralAmount;
     }
 
-    struct Action {        
-        address callee;         // eg. swapUni
-        bytes data;             // any custom param to send to the callee
-        address fromToken;      // token to be used to initiate the action   
-        address toToken;        // token to be gotten as output
+    struct Action {
+        address callee; // eg. swapUni
+        bytes data; // any custom param to send to the callee
+        address fromToken; // token to be used to initiate the action
+        address toToken; // token to be gotten as output
     }
 
-    enum Ops { GT, LT }
+    enum Ops {
+        GT,
+        LT
+    }
 
     // If Trigger and RETypes.Action are update, the HashRule needs to be updated
     struct Trigger {
-        address callee; 
-        bytes param;   //eg. abi.encode(["string", "string"], ["eth", "uni"]) 
-        uint value;     //eg. 1000
-        Ops op;         //eg. GT
+        address callee;
+        bytes param; //eg. abi.encode(["string", "string"], ["eth", "uni"])
+        uint256 value; //eg. 1000
+        Ops op; //eg. GT
     }
 }
